Display logical ESP warning in the Manage flags dialog (#291)
Finish this off and display the UEFI firmware doesn't support booting
from a logical EFI System Partition (ESP) when necessary.  If there's no
warning the widgets are hidden and there is empty space in the dialog
instead.

Add this extra code and complexity so that a user receives immediate
feedback when setting the ESP flag on a logical partition that it's not
supported for UEFI mode booting.  Without this enhancement the user has
to close the Manage flags dialog and after the device refresh know to
open the partition Information about dialog to see the warning.  If
there was already another warning for the partition then the warning
symbol wouldn't be new either.

Closes #291 - Warn about EFI System set on a logical partition