Prevent Manage flags dialog resizing when warning is shown vs hidden (#291)
Toggle the ESP flag on and off on a MSDOS logical partition.  As the
warning message is shown and hidden the dialog changes it's width.  The
amount the dialog resizes depends on the size of the font the desktop
environment uses.  The larger the font the wider the dialog becomes when
the warning message is displayed.  The largest font encountered was on
Fedora 42 with GNOME 48 using Adwaita Sans 11 [1].

A Gtk::Label widget defaults max_width_chars to 40.  GParted enables
line wrap on the m_warning_message label widget.  The logical ESP
warning message wraps on to 2 lines.  Reserving space for 3 lines of
text for translation.  On Fedora 42 with GNOME 48 using Adwaita Sans 11
font the label widget becomes 345 x 57 pixels.  Add border and frame
results in a size of 357 x 88 pixels.  Reserve 360 x 90 pixels for the
warning frame and message.  This stops the dialog resizing when the
warning message switches between being shown and hidden.

[1] GNOME 48 Switches Over To "Adwaita Sans" As Default Font
    https://www.phoronix.com/news/GNOME-48-Adwaita-Sans
        After a decade and a half of using Cantarell as the default
        font, GNOME 48 shifts over to using Adwaita Sans by default.
        ...
        schemas/org.gnome.desktop.interface.gschema.xml.in
         105 105        </description>
         106 106      </key>
         107 107      <key name="font-name" type="a">
         108     -      <default>'Cantarell 11'</default>
             108 +      <default>'Adwaita Sans 11'</default>
         109 109        <summary>Default font</summary>
         110 110        <description>
         111 111          Name of the default font used by gtk+.

Closes #291 - Warn about EFI System set on a logical partition