Update widget border widths to 5 in the File System Support dialog
The File System Support dialog uses widget border widths of 6 pixels,
where as the majority of the dialogs in GParted uses a border width of 5
pixels.
    $ egrep -oh 'set_border_width\( *[0-9]+ *\)' src/*.cc | tr -d ' ' | sort | uniq -c | sort -n
          3 set_border_width(6)
          5 set_border_width(10)
         11 set_border_width(5)

When the File System Support dialog was first updated to set the border
width [1] it was set to 6 pixels.  However at the time the majority of
code also used 5 pixels.
    $ git checkout b84a88b5579c25ed7a5e05c20f58ea044a6897d3~1
    HEAD is now at fc801411 some small changes to make the GUI more consistent (#353140)
    $ egrep -oh 'set_border_width\( *[0-9]+ *\)' src/*.cc | tr -d ' ' | sort | uniq -c | sort -n
          5 set_border_width(10)
          6 set_border_width(5)

There's no documented reason why the border width was set to 6 pixels
for this dialog.  The difference of 1 pixel in various borders is
impossible to see visually.  Therefore change it to the current
predominate value of 5 pixels.

[1] b84a88b5579c25ed7a5e05c20f58ea044a6897d3
    higgiefied the dialog a bit (#353104)