Make Dialog_Progress::m_fraction const and use std::min()
m_fraction is never updated after initialisation in the constructor so
make it const.  Update the order in the initialisation list to match the
new order in the class definition as members are initialised in order of
declaration in the class [1][2].

Replace ternary conditional operator (?:) with std::min() to avoid
repetition of multiplication in two of the operands.

[1] Constructors and member initializer lists
    https://en.cppreference.com/w/cpp/language/initializer_list.html
        Initialization order
        ...
        3) Then, non-static data member are initialized in order of
           declaration in the class definition.

[2]
    https://isocpp.org/wiki/faq/ctors#ctor-initializer-order
        Immediate base classes (left to right), then member objects (top
        to bottom).

        In other words, the order of the initialization list should
        mimic the order in which initializations will take place.  This
        guideline discourages a particularly subtle class of order
        dependency errors ...