Don't attempt to convert ext4 to 64-bit when already 64-bit (#295)
After applying the previous commit and successfully converting the ext4
file system to 64-bit and growing it to 16.01 TiB, try to grow the file
system again.  This appears to succeed:
    Grow /dev/loop0p1 from 16.01 TiB to 16.99 TiB              (SUCCESS)
    * calibrate /dev/loop0p1                                   (SUCCESS)
    * check file system on /dev/loop0p1 for errors and (if ... (SUCCESS)
    * grow partition from 16.01 TiB to 16.99 TiB               (SUCCESS)
    * grow file system to fill the partition                   (SUCCESS)
      * resize2fs -p -b '/dev/loop0p1'                         (SUCCESS)
          resize2fs 1.45.6 (20-Mar-2020)
          The filesystem is already 64-bit.

But on refresh it is clear the partition was grown but the file system
was not because there was nearly 1 TiB of unallocated space.  Resize2fs
wrote message "The filesystem is already 64-bit" to stderr, but returned
exit status 0.

Fix this by querying the ext4 file system features and only adding the
-b (convert to 64-bit) option when the file system is not already
64-bit.

Closes #295 - GParted should not disable 64bit feature of ext4 file
              systems it creates