Update main window screenshot in the GParted Manual (!152)
The screenshots on the website (https://gparted.org/screenshots.php)
were recently updated using the latest GParted application version and
to have a consistent look by all having been taken on the same desktop
environment.  Now update the main window screenshots in the GParted
Manual and its translations to match.

The screenshots were taken on a Fedora 42 VM with XFCE desktop.  The
default application font and window title font were set to Cantarell
Regular 11 and Cantarell Bold 11 respectively.  (This font was chosen
for clarity and because it matches the font used in GNOME 3 and GNOME 4,
before release 48).

Run GParted in each language and capture the screenshots thus:
    ...
    $ sudo make install
    $ LANGUAGE=C sudo gparted
Take screenshot and quit app.
(The default main window size was 775 x 500 pixels, 785 x 534 with
window manager decoration).

    $ LANGUAGE=el sudo gparted
If necessary, make the GParted window wide enough to show all columns.
The Greek translation had the longest column heading names and needed
the widest main window.  Take screenshot and quit app.
(Main window screenshot in Greek with window manager decoration was
905 x 534 pixels).

Repeat for each translation of the GParted Manual with a screenshot.
    $ ls -l help/*/figures/gparted_window.png

Note that the Indonesian translation explicitly (committed in git [1])
has it's screenshot symlinked to the C locale copy, so the Indonesian
translation of the GParted Manual displayed the English screenshot.
    $ ls -l help/id/figures/gparted_window.png
    lrwxrwxrwx. ... help/id/figures/gparted_window.png -> ../../C/figures/gparted_window.png

The GParted application and GParted Manual are both well translated to
Indonesian.  Therefore also provide an updated Indonesian screenshot for
the Manual.

The GParted Manual has a comment [2] that the GNOME Document Style Guide
[3] recommends a maximum image width of 510 pixels.  However that
document no longer appears to exist on the GNOME web site.  The nearest
found current document was the Developer Documentation Style Guidelines
[4] but that doesn't mentions anything about screenshots.

Having a screenshot as close to it's original size as possible makes the
text easier to read.  Therefore determine the maximum image size
ourselves.  Use the Greek screenshot, as it's the widest one, so all
screenshots are resized by the same percentage for consistency.  Copy
the image into the installed GParted Manual:
    $ cd ~/Pictures
    $ sudo cp gparted_window.el.png /usr/local/share/help/C/gparted/figures/gparted_window.png
    $ sudo cp gparted_window.el.png /usr/local/share/help/el/gparted/figures/gparted_window.png

And view the English and Greek manuals:
    $ LANGUAGE=C yelp help:gparted
    $ LANGUAGE=el yelp help:gparted

With trial and error found that the largest possible size for the Greek
screenshot was 70% (width 634 pixels).  Larger than this then when yelp
rendered the page for printing to A4 it had to scale down the contents
to make the image fit, which also shrank the Manual's text.  Choose A4
paper (width 210 mm) as it's slightly narrower that US letter (width 8.5
inches, ~216 mm).  Viewed on a screen the yelp window can be narrowed to
800 pixels and the GParted screenshot is still fully displayed.

Resize all the screenshots in place to 70% using this GraphicsMagick
command:
    $ cd ~/Pictures
    $ gm mogrify -filter Sinc -resize 70% gparted_window.*.png

Evaluated all the filters GraphicsMagick provides and found that Sinc
provided the clearest text in the shrunk screenshots.

This commit just updates the C locale screenshot and the comment in the
Manual with the discovered maximum image width.

[1] commit f250697c58268c92e8c3bff74145c23bb6133ed5
    Add initial Indonesian translation of help (!90)

[2] commit 8bc1377c2232934eb50bcb84454560312b9acbc9
    Add comment for documentation maximum image width recommendation

[3] GNOME Documentation Style Guide V1.6, 10.3 Image Width
    http://web.archive.org/web/20210306042742/https://developer.gnome.org/gdp-style-guide/stable/screenshot-4.html.en
        The maximum image width is 510 pixels.  At this width, images
        fit on a printed page.  You might need to scale down the image
        before you add any callouts, to ensure that the total width of
        the image is less than 510 pixels.  You can use GIMP to scale
        images before you add callouts.  To scale an image in GIMP,
        right-click on the image, then choose Image > Scale image.

[4] Developer Documentation Style Guidelines
    https://developer.gnome.org/documentation/guidelines/devel-docs.html

Closes !152 - Update main window screenshot in the GParted Manual and
              its translations