Update German screenshot for the translated Manual (!152)
Note that the German translation PO (Portable Object) file uses a
different (older?) style of referencing the external image [1] compared
to that used in the Czech translation.  As a result the GNOME website
marks the image as fuzzy instead of translated [2][3][4].

All these GParted Manual translations use the same "@@image:" style of
referencing the external image and all report the translation of the
image as fuzzy whether the image exists or not!
    Code       Language          Image           Translation
    de         German            exists          fuzzy
    el         Greek             exists          fuzzy
    gd         Scottish Gaelic   doesn't exist   fuzzy
    it         Italian           exists          fuzzy
    sl         Slovenian         doesn't exist   fuzzy
    sr         Serbian           exists          fuzzy
    sr@latin   Serbian (Latin)   exists          fuzzy
    te         Telugu            doesn't exist   fuzzy

Getting the [Current translated file (PO)] from the gparted > master >
German page [2] provides a reformatted file that looks like it would
remove the fuzzy status of the image translation.  However it reformats
the whole file generating this level of change:
    1857 insertions(+), 915 deletions(-)
Changes include updating itstool comments, re-ordering translations and
re-line wrapping source and translated text.  Expect that this shouldn't
affect the translation but it seems like something that should be left
to the translation teams to perform and verify.

So just update German screenshot and the md5sum hash of the C locale
screenshot in the German PO file.

[1] GParted German PO file
    help/de/de.po
        #. When image changes, this message will be marked fuzzy or
        #. untranslated for you. It doesn't matter what you translate it
        #. to: it's not used at all.
        #: C/gparted.xml:382(None)
        msgid ""
        "@@image: 'figures/gparted_window.png'; md5=0c6eb27cd7d5d614b4ea752846ccb50b"
        msgstr "translated"

[2] gparted > master > German
    https://l10n.gnome.org/vertimus/gparted/master/help/de/
        including 0/1/0 image   [Translated/Fuzzy/Untranslated]

[3] User Guide (German), Figures translation status
    https://l10n.gnome.org/module/gparted/help/master/de/images/

[4] gparted > master > German > Differences for fuzzy strings
    https://l10n.gnome.org/vertimus/stats/31689/msgiddiff/
        512#, fuzzy
            -@@image: 'figures/gparted_window.png'; md5=0c6eb27cd7d5d614b4ea752846ccb50b
            +msgctxt "_external ref='figures/gparted_window.png' md5='0c6eb27cd7d5d614b4ea752846ccb50b'
        525msgctxt "_"
        526msgid ""
        527"external ref='figures/gparted_window.png' "
        528"md5='0c6eb27cd7d5d614b4ea752846ccb50b'"
        529msgstr "translated"

Closes !152 - Update main window screenshot in the GParted Manual and
              its translations