Add Indonesian screenshot for the translated Manual (!152)
As noted earlier the Indonesian screenshot was a symbolic link to the C
locale screenshot.
    $ ls -l help/id/figures/gparted_window.png
    lrwxrwxrwx. ... help/id/figures/gparted_window.png -> ../../C/figures/gparted_window.png

This displayed the English screenshot in the Indonesian GParted Manual
and caused a warning against the translation in the GNOME translation
website [1][2].  The GParted application and GParted Manual are well
translated to Indonesian, 100% and 98% respectively.  Therefore provide
an Indonesian screenshot.

[1] The GNOME Project Translation Platform > Modules > gparted
    https://l10n.gnome.org/module/gparted/

[2] gparted > master > Indonesian [User Guide]
    https://l10n.gnome.org/vertimus/gparted/master/help/id/
        /!\ Figures should not be copied when identical to original
        (/var/www/damnedlies/data/scratchdir/git/gparted/help/id/figures/gparted_window.png).

Closes !152 - Update main window screenshot in the GParted Manual and
              its translations