Set LBA flag when pasting FAT16/32 into new MSDOS partition (#300)
Copy an existing FAT16/32 partition without the LBA flag set on an MSDOS
drive into a new partition.  The operation previews and implements not
setting the LBA flag.  Where as copying the same partition into an
existing partition previews and implements as setting the LBA flag.

Fix this so copying into a new partition also sets the partition flags
consistently so when previewing and implementing this operation the LBA
flag is set.

Closes #300 - Consider setting LBA flag for FAT32 primary partitions
              automatically when needed