Remove unnecessary comment from Dialog_Partition_Copy::get_new_partition()
The comment and two associated lines of code were added by commit [1] on
2004-09-25.  5 commits later [2] on the same day the first commented
statement was removed.  So now the first clause of the comment is not
applicable any more and the second clause "set status for partition" is
just repeating exactly what can be seen from the remaining statement
itself and adds no additional information.  Therefore remove.

[1] commit 201fdcb8e8ee8587775b79544ee3ff4bb87d8cd6
    more i18n and some code cleanups
            //set proper name and status for partition
        //  selected_partition.partition = String::ucompose( _("copy from %1"), selected_partition .partition );
            selected_partition.status = GParted::STAT_COPY ;

[2] commit 60b5bb29aad0c29c78df8784c61c8961eb624a8e
    removed commented statement (big deal huh:) )