Clear all flags when setting partition type by fstype (#300)
Here are several examples of the same issue.  Have an existing partition
on the relevant table with the specified libparted flag set and apply
the operation to that partition.
    Table   Partition num, FS, flag     Apply operation
    MSDOS   1   unknown      raid       Format to ext4
    GPT     1   linux-swap   swap       Paste fat16 into
    GPT     2   fat16        msftdata   Format to swap

In each case after refresh the existing partition flag was maintained
even though the the type of the new file system was set using libparted.
This is because the libparted flags are overriding the request being
made by setting the partition's file system type in setting the on-disk
partition type.

So far set_filesystem_type_using_fstype() only clears the ESP and LVM
flags.  Fix this by clearing all libparted partition flags when setting
the on-disk partition type by file system type.

Closes #300 - Consider setting LBA flag for FAT32 primary partitions
              automatically when needed